performance improvements for monitorBonds

+++ src/commands/monitor_bonds.py
@@ -21,11 +21,17 @@ from chimerax.core.commands import (
 
 from AaronTools.const import RADII, METAL
 
+# import os
+# os.environ["LINE_PROFILE"] = "1"
+
+# import line_profiler
+
 max_dist_arrays = {}
 ts_dist_arrays = {}
 tracked_atoms = {}
 handlers = {}
 
+# @line_profiler.profile
 def check_bonds(trigger_name, changes, tolerance, tsTolerance):
     """
     if the distance between two atoms is shorter than the sum of
@@ -39,8 +45,7 @@ def check_bonds(trigger_name, changes, tolerance, tsTolerance):
     if not diff.structure_reasons():
         return
     # squared distance between atoms
-    D = distance.squareform(
-        distance.pdist(
+    D = distance.squareform(distance.pdist(
             mdl.coordset(mdl.active_coordset_id).xyzs[tracked_atoms[mdl], :],
             "sqeuclidean"
         )
@@ -51,60 +56,46 @@ def check_bonds(trigger_name, changes, tolerance, tsTolerance):
     except KeyError:
         # make an array for how far atoms can be and still be bonded
         # to each other
-        max_connected_dist = np.zeros((len(tracked_atoms[mdl]), len(tracked_atoms[mdl])))
-        ts_connected_dist = np.zeros((len(tracked_atoms[mdl]), len(tracked_atoms[mdl])))
-        for i, ii in enumerate(tracked_atoms[mdl]):
-            a1 = mdl.atoms[ii]
-            for j, jj in enumerate(tracked_atoms[mdl][:i]):
-                a2 = mdl.atoms[jj]
-                max_connected_dist[i, j] = (
-                    RADII[a1.element.name] +
-                    RADII[a2.element.name] +
-                    tolerance
-                ) ** 2
-                ts_connected_dist[i, j] = (
-                    RADII[a1.element.name] +
-                    RADII[a2.element.name] +
-                    tsTolerance
-                ) ** 2
+        radii_list = np.vectorize(
+            lambda i: RADII.get(mdl.atoms[i].element.name, 0)
+        )(tracked_atoms[mdl])
+        max_connected_dist = np.add.outer(radii_list, radii_list) + tolerance
+        max_connected_dist = max_connected_dist ** 2
+        ts_connected_dist = np.add.outer(radii_list, radii_list) + tsTolerance
+        ts_connected_dist = ts_connected_dist ** 2
+
         max_dist_arrays[mdl] = max_connected_dist
         ts_dist_arrays[mdl] = ts_connected_dist
 
     # these are bonded
-    connected = np.tril(D < max_connected_dist)
+    connected = np.tril(D < max_connected_dist, k=-1)
     # these are not
-    not_connected = np.tril(D > max_connected_dist)
+    not_connected = np.tril(D > max_connected_dist, k=-1)
     # almost bonded
-    almost_connected = np.tril(np.logical_and(not_connected, D < ts_connected_dist))
+    almost_connected = np.tril(np.logical_and(not_connected, D < ts_connected_dist), k=-1)
     # way too long
-    way_too_long = np.tril(D > ts_connected_dist)
+    way_too_long = np.tril(D > ts_connected_dist, k=-1)
 
+    ts_pbg = mdl.pseudobond_group(
+        "TS bonds",
+        create_type=0,
+    )
     # add new bonds
-    for (ndx1, ndx2) in zip(*connected.nonzero()):
-        a1 = mdl.atoms[tracked_atoms[mdl][ndx1]]
-        a2 = mdl.atoms[tracked_atoms[mdl][ndx2]]
-        is_metal = False
-        if (
-            any(a1.element.name == ele for ele in METAL)
-        ) or (
-            any(a2.element.name == ele for ele in METAL)
-        ):
-            is_metal = True
-        if not is_metal:
+    n_metal_bonds_counted = 0
+    bonded_pairs = connected.nonzero()
+    g1 = mdl.atoms[tracked_atoms[mdl][bonded_pairs[0]]]
+    g2 = mdl.atoms[tracked_atoms[mdl][bonded_pairs[1]]]
+    for (a1, a2) in zip(g1, g2):
+        if a1 in a2.neighbors:
+            continue
+        if (a1.element.name not in METAL) and (a2.element.name not in METAL):
             try:
                 b = mdl.new_bond(a1, a2)
                 # set the bond radius to match other bonds
-                radius = 0
-                n_bonds = 0
-                for bond in a1.bonds:
-                    if bond is b:
-                        continue
-                    radius += bond.radius
-                    n_bonds += 1
-                if radius == 0:
-                    b.radius = 0.16
-                else:
-                    b.radius = (radius / n_bonds)
+                radius = np.mean(mdl.bonds.radii)
+                if np.isnan(radius):
+                    radius = 0.16
+                b.radius = radius
             except TypeError:
                 pass
         else:
@@ -120,14 +111,12 @@ def check_bonds(trigger_name, changes, tolerance, tsTolerance):
                 )
             for pbond in pbg.pseudobonds:
                 if a1 in pbond.atoms and a2 in pbond.atoms:
+                    n_metal_bonds_counted += 1
                     break
             else:
                 pbg.new_pseudobond(a1, a2)
+                n_metal_bonds_counted += 1
 
-        ts_pbg = mdl.pseudobond_group(
-            "TS bonds",
-            create_type=0,
-        )
         if ts_pbg is None:
             continue
         for pbond in ts_pbg.pseudobonds:
@@ -136,54 +125,65 @@ def check_bonds(trigger_name, changes, tolerance, tsTolerance):
                 break
 
     # delete bonds that are too long
-    for (ndx1, ndx2) in zip(*not_connected.nonzero()):
-        a1 = mdl.atoms[tracked_atoms[mdl][ndx1]]
-        a2 = mdl.atoms[tracked_atoms[mdl][ndx2]]
-        for bond in a1.bonds:
-            if a2 in bond.atoms:
-                bond.delete()
-                break
-        pbg = mdl.pseudobond_group(
-            mdl.PBG_METAL_COORDINATION,
-            create_type=0,
-        )
-        if pbg is None:
-            continue
-        for pbond in pbg.pseudobonds:
-            if a1 in pbond.atoms and a2 in pbond.atoms:
-                pbond.delete()
-                break
+    pbg = mdl.pseudobond_group(
+        mdl.PBG_METAL_COORDINATION,
+        create_type=0,
+    )
+    no_bonds_pairs = not_connected.nonzero()
+    g1 = mdl.atoms[tracked_atoms[mdl][no_bonds_pairs[0]]]
+    g2 = mdl.atoms[tracked_atoms[mdl][no_bonds_pairs[1]]]
+    for (a1, a2) in zip(g1, g2):
+        if a1 in a2.neighbors:
+            for bond in a1.bonds:
+                if a2 in bond.atoms:
+                    bond.delete()
+                    break
+        if (a1.element.name in METAL) or (a2.element.name in METAL):
+            if pbg is None:
+                continue
+            if pbg.num_pseudobonds - n_metal_bonds_counted == 0:
+                continue
+            for pbond in pbg.pseudobonds:
+                if a1 in pbond.atoms and a2 in pbond.atoms:
+                    pbond.delete()
+                    break
 
-    for (ndx1, ndx2) in zip(*almost_connected.nonzero()):
-        a1 = mdl.atoms[tracked_atoms[mdl][ndx1]]
-        a2 = mdl.atoms[tracked_atoms[mdl][ndx2]]
+    n_ts_bonds_counted = 0
+    ts_bonds_pairs = almost_connected.nonzero()
+    g1 = mdl.atoms[tracked_atoms[mdl][ts_bonds_pairs[0]]]
+    g2 = mdl.atoms[tracked_atoms[mdl][ts_bonds_pairs[1]]]
+    for (a1, a2) in zip(g1, g2):
         ts_pbg = mdl.pseudobond_group(
             "TS bonds",
             create_type=1,
         )
         for pbond in ts_pbg.pseudobonds:
             if a1 in pbond.atoms and a2 in pbond.atoms:
+                n_ts_bonds_counted += 1
                 break
         else:
             ts_pbg.new_pseudobond(a1, a2)
+            n_ts_bonds_counted += 1
         ts_pbg.dashes = 0
         ts_pbg.halfbond = False
         ts_pbg.color = [170, 255, 255, 128]
         ts_pbg.radius = 0.16
     
-    for (ndx1, ndx2) in zip(*way_too_long.nonzero()):
-        a1 = mdl.atoms[tracked_atoms[mdl][ndx1]]
-        a2 = mdl.atoms[tracked_atoms[mdl][ndx2]]
-        ts_pbg = mdl.pseudobond_group(
-            "TS bonds",
-            create_type=0,
-        )
-        if ts_pbg is None:
-            break
-        for pbond in ts_pbg.pseudobonds:
-            if a1 in pbond.atoms and a2 in pbond.atoms:
-                pbond.delete()
+    ts_pbg = mdl.pseudobond_group(
+        "TS bonds",
+        create_type=0,
+    )
+    if ts_pbg is not None:
+        no_ts_bonds = way_too_long.nonzero()
+        g1 = mdl.atoms[tracked_atoms[mdl][no_ts_bonds[0]]]
+        g2 = mdl.atoms[tracked_atoms[mdl][no_ts_bonds[1]]]
+        for (a1, a2) in zip(g1, g2):
+            if ts_pbg.num_pseudobonds - n_ts_bonds_counted == 0:
                 break
+            for pbond in ts_pbg.pseudobonds:
+                if a1 in pbond.atoms and a2 in pbond.atoms:
+                    pbond.delete()
+                    break
 
 def monitor_bonds(session, selection, tolerance=0.35, tsTolerance=0.6, guess=False):
     """
@@ -221,8 +221,7 @@ def monitor_bonds(session, selection, tolerance=0.35, tsTolerance=0.6, guess=Fal
                         bonds_changed.update((i, j))
             atoms = Atoms([atoms[b] for b in bonds_changed])
         ndx = [mdl.atoms.index(a) for a in atoms]
-        tracked_atoms.setdefault(mdl, list())
-        tracked_atoms[mdl].extend(ndx)
+        tracked_atoms[mdl] = np.array(ndx, dtype=int)
         hdlr = mdl.triggers.add_handler(
             "changes",
             lambda tn, changes, tolerance=tolerance, tsTolerance=tsTolerance: \
